(PC-19327) feat(favorites): extract getFavoriteDisplayPrice and move to Favorite feature
