+++ src/features/favorites/components/Favorite.tsx
@@ -6,11 +6,12 @@ import styled, { useTheme } from 'styled-components/native'
 import { FavoriteOfferResponse, FavoriteResponse, UserProfileResponse } from 'api/gen'
 import { useRemoveFavorite } from 'features/favorites/api'
 import { BookingButton } from 'features/favorites/components/Buttons/BookingButton'
+import { getFavoriteDisplayPrice } from 'features/favorites/helpers/getFavoriteDisplayPrice'
 import { mergeOfferData } from 'features/offer/components/OfferTile/OfferTile'
 import { useShareOffer } from 'features/offer/helpers/useShareOffer'
 import { analytics } from 'libs/firebase/analytics'
 import { useDistance } from 'libs/geolocation/hooks/useDistance'
-import { formatToFrenchDate, getFavoriteDisplayPrice } from 'libs/parsers'
+import { formatToFrenchDate } from 'libs/parsers'
 import { QueryKeys } from 'libs/queryKeys'
 import { WebShareModal } from 'libs/share/WebShareModal'
 import { useSearchGroupLabel, useSubcategory } from 'libs/subcategories'
