+++ src/libs/parsers/getDisplayPrice.ts
@@ -35,18 +35,3 @@ export const getDisplayPriceWithDuoMention = (prices: number[] | undefined): str
   if (prices.includes(0)) return 'Gratuit'
   return `${getPricePerPlace(prices)} / place`
 }
-
-export const getFavoriteDisplayPrice = ({
-  startPrice,
-  price,
-}: {
-  startPrice?: number | null
-  price?: number | null
-}): string => {
-  if (price === 0) return 'Gratuit'
-  if (price && price > 0) return formatToFrenchDecimal(price)
-  if (startPrice === 0 || (startPrice && startPrice > 0)) {
-    return `Dès ${formatToFrenchDecimal(startPrice)}`
-  }
-  return ''
-}
