+++ src/libs/parsers/tests/getDisplayPrice.test.ts
@@ -1,7 +1,6 @@
 import {
   getDisplayPrice,
   getDisplayPriceWithDuoMention,
-  getFavoriteDisplayPrice,
   formatToFrenchDecimal,
   formatPriceInEuroToDisplayPrice,
 } from '../getDisplayPrice'
@@ -41,28 +40,6 @@ describe('getDisplayPrice', () => {
   `('getDisplayPriceWithDuoMention($prices) \t= $expected', ({ prices, expected }) => {
     expect(getDisplayPriceWithDuoMention(prices)).toBe(expected)
   })
-})
-
-describe('getFavoriteDisplayPrice', () => {
-  it.each`
-    price        | startPrice   | expected
-    ${undefined} | ${undefined} | ${''}
-    ${null}      | ${null}      | ${''}
-    ${0}         | ${null}      | ${'Gratuit'}
-    ${1000}      | ${null}      | ${'10\u00a0€'}
-    ${null}      | ${0}         | ${'Dès 0\u00a0€'}
-    ${null}      | ${1000}      | ${'Dès 10\u00a0€'}
-  `(
-    'getFavoriteDisplayPrice({ price: $price, startPrice: $startPrice }) \t= $expected',
-    ({ price, startPrice, expected }) => {
-      expect(
-        getFavoriteDisplayPrice({
-          price,
-          startPrice,
-        })
-      ).toBe(expected)
-    }
-  )
 })
 
 describe('formatToFrenchDecimal()', () => {
